create router for app
