+++ exercise_react/package.json
@@ -45,9 +45,11 @@
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
+    "@types/react-router-dom": "^5.1.7",
     "axios": "^0.21.1",
     "moment": "^2.29.1",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0"
   }
 }
