+++ exercise_react/src/components/Post.tsx
