+++ exercise_react/src/components/PostDetail.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import '../style/style.scss';
+
+const PostDetail = () => {
+  return (
+    <>
+      <h3>Hello</h3>
+    </>
+  )
+}
+export default PostDetail;
