+++ exercise_react/src/style/base/_reset.scss
@@ -20,6 +20,10 @@ li {
   list-style: none;
 }
 
+a {
+  text-decoration:  none;
+}
+
 select {
   appearance: none;
 }
